alx_travel_app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1